+++ src/main/java/com/dexma/hometest/business/ChangeProcessor.java
@@ -1,5 +1,6 @@
 package com.dexma.hometest.business;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import com.dexma.hometest.domain.Cash;
@@ -10,5 +11,5 @@ import com.dexma.hometest.domain.Cash;
  */
 public interface ChangeProcessor
 {
-    Map<Cash, Integer> processChange(final Map<Cash, Integer> cashStock, double amount);
+    Map<Cash, Integer> processChange(final Map<Cash, Integer> cashStock, BigDecimal amount);
 }
