+++ src/main/java/com/dexma/hometest/domain/Product.java
@@ -1,5 +1,8 @@
 package com.dexma.hometest.domain;
 
+import java.math.BigDecimal;
+
+
 /**
  *
  */
@@ -7,5 +10,5 @@ public interface Product
 {
     String getName();
 
-    double getPrice();
+    BigDecimal getPrice();
 }
